codebase_genius/backend/requirements.txt
jaclang
GitPython
tree_sitter
graphviz
markdown
requests